+++ CHANGELOG.md
@@ -67,4 +67,8 @@ January 15th, 2024
 - Updated instance configuration page.
     - Removed the frame-rate configuration option to reflect Predator's updated configuration layout.
     - Added more resolution options.
+    - Added more GPS video overlay stamp configuration options.
+    - Added parking mode configuration options.
 - Updated the storage page to be compatible with Predator V10.0.
+    - Optic is now compatible with the new Predator configuration layout.
+    - Optic now recognizes separate audio files and displays them next to their associated video file.